+++ src/@types/global.d.ts
@@ -1,0 +1,12 @@
+declare global {
+  namespace NodeJS {
+    interface Global {
+      document: Document;
+      window: Window;
+      navigator: Navigator;
+    }
+  }
+  interface Window {
+    IntersectionObserver: IntersectionObserver;
+  }
+}