+++ src/components/components/Navbar/NavSection.ts
@@ -4,7 +4,6 @@ export const NavSection = styled.section`
   align-items: center;
   display: flex;
   flex: 1 0 0;
-
   &:not(:first-child):last-child {
     justify-content: flex-end;
   }