+++ src/components/elements/Picture/types.d.ts
@@ -1,0 +1,12 @@
+import { GlobProps } from "../../../typings";
+
+type srcSetProps = {
+  media: string;
+  src: string;
+};
+
+export interface PictureProps extends GlobProps {
+  dataSrc?: string;
+  alt?: string;
+  srcSets?: srcSetProps[];
+}