+++ src/components/layout/Flexbox/index.test.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { render, cleanup } from "@testing-library/react";
 import Flexbox from "./index";
+import Provider from "../../layout/Provider";
 
 afterAll(cleanup);
 
@@ -8,20 +9,30 @@ describe("<Flex/>", () => {
   // Render test
   it("should renders correctly", () => {
     const { container } = render(
-      <Flexbox>
-        <Flexbox.Column />
-      </Flexbox>
+      <Provider>
+        <Flexbox>
+          <Flexbox.Column />
+        </Flexbox>
+      </Provider>
     );
     expect(container).toMatchSnapshot();
   });
 
   it("should renders gapless", () => {
-    const { container } = render(<Flexbox gapless />);
+    const { container } = render(
+      <Provider>
+        <Flexbox gapless />
+      </Provider>
+    );
     expect(container).toMatchSnapshot();
   });
 
   it("should renders correctly", () => {
-    const { container } = render(<Flexbox oneline />);
+    const { container } = render(
+      <Provider>
+        <Flexbox oneline />
+      </Provider>
+    );
     expect(container).toMatchSnapshot();
   });
 });